+++ tests/base/test_multitenancy.py
@@ -48,15 +48,7 @@ def app(component_builder):
 
 @pytest.mark.parametrize("response_test", [
     ResponseTest(
-            "http://dummy_uri/parse?q=food&project=test_project_mitie",
-            {"entities": [], "intent": "affirm", "text": "food"}
-    ),
-    ResponseTest(
-            "http://dummy_uri/parse?q=food&project=test_project_mitie_sklearn",
-            {"entities": [], "intent": "restaurant_search", "text": "food"}
-    ),
-    ResponseTest(
-            "http://dummy_uri/parse?q=food&project=test_project_spacy_sklearn",
+            "http://dummy-uri/parse?q=food&project=test_project_spacy_sklearn",
             {"entities": [], "intent": "restaurant_search", "text": "food"}
     ),
 ])
@@ -71,11 +63,11 @@ def test_get_parse(app, response_test):
 
 @pytest.mark.parametrize("response_test", [
     ResponseTest(
-            "http://dummy_uri/parse?q=food",
+            "http://dummy-uri/parse?q=food",
             {"error": "No project found with name 'default'."}
     ),
     ResponseTest(
-            "http://dummy_uri/parse?q=food&project=umpalumpa",
+            "http://dummy-uri/parse?q=food&project=umpalumpa",
             {"error": "No project found with name 'umpalumpa'."}
     )
 ])
@@ -89,17 +81,7 @@ def test_get_parse_invalid_model(app, response_test):
 
 @pytest.mark.parametrize("response_test", [
     ResponseTest(
-            "http://dummy_uri/parse",
-            {"entities": [], "intent": "affirm", "text": "food"},
-            payload={"q": "food", "project": "test_project_mitie"}
-    ),
-    ResponseTest(
-            "http://dummy_uri/parse",
-            {"entities": [], "intent": "restaurant_search", "text": "food"},
-            payload={"q": "food", "project": "test_project_mitie_sklearn"}
-    ),
-    ResponseTest(
-            "http://dummy_uri/parse",
+            "http://dummy-uri/parse",
             {"entities": [], "intent": "restaurant_search", "text": "food"},
             payload={"q": "food", "project": "test_project_spacy_sklearn"}
     ),
@@ -114,23 +96,27 @@ def test_post_parse(app, response_test):
 
 @pytest.inlineCallbacks
 def test_post_parse_specific_model(app):
-    status = yield app.get("http://dummy_uri/status")
+    status = yield app.get("http://dummy-uri/status")
     sjs = yield status.json()
-    model = sjs["available_projects"]["test_project_mitie"]["available_models"][0]
-    query = ResponseTest("http://dummy_uri/parse", {"entities": [], "intent": "affirm", "text": "food"},
-                         payload={"q": "food", "project": "test_project_mitie", "model": model})
+    project = sjs["available_projects"]["test_project_spacy_sklearn"]
+    model = project["available_models"][0]
+    query = ResponseTest("http://dummy-uri/parse",
+                         {"entities": [], "intent": "affirm", "text": "food"},
+                         payload={"q": "food",
+                                  "project": "test_project_spacy_sklearn",
+                                  "model": model})
     response = yield app.post(query.endpoint, json=query.payload)
     assert response.code == 200
 
 
 @pytest.mark.parametrize("response_test", [
     ResponseTest(
-            "http://dummy_uri/parse",
+            "http://dummy-uri/parse",
             {"error": "No project found with name 'default'."},
             payload={"q": "food"}
     ),
     ResponseTest(
-            "http://dummy_uri/parse",
+            "http://dummy-uri/parse",
             {"error": "No project found with name 'umpalumpa'."},
             payload={"q": "food", "project": "umpalumpa"}
     ),
@@ -157,6 +143,4 @@ def train_models(component_builder):
         persistor = create_persistor(config)
         trainer.persist("test_projects", persistor, project_name)
 
-    train("sample_configs/config_mitie.json", "test_project_mitie")
     train("sample_configs/config_spacy.json", "test_project_spacy_sklearn")
-    train("sample_configs/config_mitie_sklearn.json", "test_project_mitie_sklearn")